+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -106,8 +106,8 @@ New checks
   Checks for uses of MLIR's old/to be deprecated ``OpBuilder::create<T>`` form
   and suggests using ``T::create`` instead.
 
-- New :doc:`misc-visibility-change-to-virtual-function
-  <clang-tidy/checks/misc/visibility-change-to-virtual-function>` check.
+- New :doc:`misc-override-with-different-visibility
+  <clang-tidy/checks/misc/override-with-different-visibility>` check.
 
   Finds virtual function overrides with different visibility than the function
   in the base class.